Consolidate RAMOS metrics hooks (#796)
* Consolidate RAMOS metrics hooks

* Remove filteredHourlyMetrics

* Remove filteredHourlyMetrics from RAMOSMetrics.tsx

* Fix string type
