+++ apps/dapp/src/hooks/core/types.ts
@@ -51,14 +51,6 @@ export type Metrics = {
   tvlUSD: string;
 }[];
 
-export type RAMOSMetric = {
-  templeBurned: string;
-  totalProfitUSD: string;
-  treasuryPriceIndexUSD: string;
-  templeVolume: string;
-  timestamp: string;
-};
-
 export interface SubGraphResponse<T extends object> {
   data?: T;
   errors?: { message: string }[];
@@ -72,8 +64,6 @@ export interface SubGraphQuery {
 export type GetVaultGroupsResponse = SubGraphResponse<{ vaultGroups: GraphVaultGroup[] }>;
 export type GetVaultGroupResponse = SubGraphResponse<{ vaultGroup: GraphVaultGroup }>;
 export type GetMetricsResponse = SubGraphResponse<{ metrics: Metrics }>;
-export type GetRAMOSDailyMetricsResponse = SubGraphResponse<{ metricDailySnapshots: RAMOSMetric[] }>;
-export type GetRAMOSHourlyMetricsResponse = SubGraphResponse<{ metricHourlySnapshots: RAMOSMetric[] }>;
 
 export class SubgraphError extends Error {
   constructor(public message: string, public cause: Error) {
